+++ src/Database/QueryBuilder/SqliteQueryBuilder.php
@@ -77,46 +77,12 @@ class SqliteQueryBuilder extends CommonQueryBuilder implements QueryBuilderInter
      */
     public function alterTable(Table $table)
     {
-        $queries = [];
-        
-        $columns = $table->getColumns();
-        if (!empty($columns)) {
-            $query = 'ALTER TABLE ' . $this->escapeString($table->getName()) . ' ';
-            $columnList = [];
-            foreach ($columns as $column) {
-                $columnList[] = 'ADD COLUMN ' . $this->createColumn($column, $table);
-            }
-            $query .= implode(',', $columnList) . ';';
-            $queries[] = $query;
-        }
-        
+        $queries = $this->addColumns($table);
         if ($table->getColumnsToChange()) {
-            if (is_null($this->adapter)) {
-                throw new PhoenixException('Missing adapter');
-            }
-            $oldColumns = $this->adapter->tableInfo($table->getName());
-            $columns = array_merge($oldColumns, $table->getColumnsToChange());
-
             $tmpTableName = '_' . $table->getName() . '_old_' . date('YmdHis');
             $queries = array_merge($queries, $this->renameTable($table, $tmpTableName));
-            
-            $newTable = new Table($table->getName());
-            $primaryColumn = false;
-            $columnNames = [];
-            foreach ($columns as $column) {
-                $columnNames[] = $column->getName();
-                if ($column->isAutoincrement()) {
-                    $primaryColumn = $column;
-                } else {
-                    $newTable->addColumn($column);
-                }
-            }
-            $newTable->addPrimary($primaryColumn);
-            $queries = array_merge($queries, $this->createTable($newTable));
-            
-            // copy data
-            $queries[] = 'INSERT INTO ' . $this->escapeString($newTable->getName()) . ' (' . implode(',', $this->escapeArray($columnNames)) . ') SELECT ' . implode(',', $this->escapeArray(array_keys($oldColumns))) . ' FROM ' . $this->escapeString($tmpTableName);
-            
+            $queries = array_merge($queries, $this->createNewTable($table, $tmpTableName));
+
             $tableToDrop = new Table($tmpTableName);
             $queries = array_merge($queries, $this->dropTable($tableToDrop));
         }
@@ -141,13 +107,9 @@ class SqliteQueryBuilder extends CommonQueryBuilder implements QueryBuilderInter
         }
         return $col;
     }
-    
-    protected function createPrimaryKey(Table $table)
+
+    protected function primaryKeyString(Table $table)
     {
-        if (empty($table->getPrimaryColumns())) {
-            return '';
-        }
-        
         $primaryKeys = [];
         foreach ($table->getPrimaryColumns() as $name) {
             $column = $table->getColumn($name);
@@ -169,6 +131,30 @@ class SqliteQueryBuilder extends CommonQueryBuilder implements QueryBuilderInter
         }
         $query = 'CREATE ' . $index->getType() . ' ' . $this->escapeString($index->getName()) . ' ON ' . $this->escapeString($table->getName()) . ' (' . implode(',', $columns) . ');';
         return $query;
+    }
+
+    private function createNewTable(Table $table, $tmpTableName)
+    {
+        if (is_null($this->adapter)) {
+            throw new PhoenixException('Missing adapter');
+        }
+        $oldColumns = $this->adapter->tableInfo($table->getName());
+        $columns = array_merge($oldColumns, $table->getColumnsToChange());
+        
+        $newTable = new Table($table->getName());
+        $columnNames = [];
+        foreach ($columns as $column) {
+            $columnNames[] = $column->getName();
+            if ($column->isAutoincrement()) {
+                $newTable->addPrimary($column);
+                continue;
+            }
+            $newTable->addColumn($column);
+        }
+        
+        $queries = $this->createTable($newTable);
+        $queries[] = 'INSERT INTO ' . $this->escapeString($newTable->getName()) . ' (' . implode(',', $this->escapeArray($columnNames)) . ') SELECT ' . implode(',', $this->escapeArray(array_keys($oldColumns))) . ' FROM ' . $this->escapeString($tmpTableName);
+        return $queries;
     }
     
     public function escapeString($string)