+++ src/composer_autoloader.php
@@ -2,8 +2,8 @@
 
 return function () {
     $files = [
-      __DIR__ . '/../../../autoload.php',  // composer dependency
-      __DIR__ . '/../vendor/autoload.php', // stand-alone package
+        __DIR__ . '/../../../autoload.php',  // composer dependency
+        __DIR__ . '/../vendor/autoload.php', // stand-alone package
     ];
     foreach ($files as $file) {
         if (is_file($file)) {